refactor(core): split layout into types, tick derivation, and pass
Extract Domain/theme/result contracts and DEFAULT_LAYOUT_THEME into
layout-types, tick/band/temporal derivation into layout-derive-ticks,
and keep layoutPass/layout/marginDelta as the measurement orchestrator.
Re-export the public surface from layout.ts so package and pipeline
import paths stay stable.

Split layout characterization tests by concern under tests/layout/
(ordinary, fixtures, band-axis, two-pass sufficiency) with shared fixtures.
