+++ packages/core/src/layout/layout.ts
@@ -24,441 +24,42 @@
  * (band) x-axes go through the measured `planBandAxis` planner, which may wrap or
  * rotate labels within the axis band (see band-guide.ts); that layout travels on
  * the guide plan (bandLabelMode/bandLabelAngle) and enriched ticks, not here.
+ *
+ * Types live in `layout-types.ts`; tick derivation in `layout-derive-ticks.ts`.
+ * This module owns measurement / margin assembly and re-exports the public surface.
  */
 
-import type { PositionScaleSpec, TemporalKind } from "@ggsvelte/spec";
-
-import type { CellValue } from "../table.js";
-import type { AxisGuidePlan } from "./temporal-guide.js";
-import { planTemporalAxis } from "./temporal-guide.js";
-import { planBandAxis, type BandAxisPlan } from "./band-guide.js";
 import type { TextMeasurer } from "./measure.js";
+import { deriveTicks, type AxisTicks, type DeriveTicksContext } from "./layout-derive-ticks.js";
 import {
-  defaultLogTickFormat,
-  defaultTickFormat,
-  linearTicks,
-  logTicks,
-  tickStep,
-} from "./ticks.js";
-import { defaultTimeTickFormat, timeTicks } from "./time.js";
+  DEFAULT_LAYOUT_THEME,
+  type LayoutInput,
+  type LayoutResult,
+  type LayoutTheme,
+  type Margins,
+  type PassResult,
+} from "./layout-types.js";
 
-export type Domain =
-  | {
-      type: "linear";
-      /** Pre-stat transform: numeric tick/grid/format dispatch keys from this. */
-      transform?: "identity" | "log10" | "sqrt";
-      min: number;
-      max: number;
-      breaks?: readonly number[];
-    }
-  | {
-      type: "time";
-      min: number;
-      max: number;
-      breaks?: readonly number[];
-      temporal?: TemporalLayoutDomainContext;
-    }
-  | {
-      type: "band";
-      categories: string[];
-      rawCategories?: readonly unknown[];
-      breaks?: readonly (number | string)[];
-      /** Present ⇒ this band axis gets the measured label planner (horizontal only). */
-      band?: BandLayoutDomainContext;
-    };
-
-export interface TemporalLayoutDomainContext {
-  aesthetic: "x" | "y";
-  panelIndex: number;
-  kind: TemporalKind;
-  config: PositionScaleSpec;
-  sourceBreaks?: readonly CellValue[];
-}
-
-export interface BandLayoutDomainContext {
-  /** Data aesthetic (flip-corrected) — used for the guide-plan id and diagnostic path. */
-  aesthetic: "x" | "y";
-  panelIndex: number;
-  config: PositionScaleSpec;
-}
-
-export interface Margins {
-  top: number;
-  right: number;
-  bottom: number;
-  left: number;
-}
-
-export interface Tick {
-  value: number | string;
-  label: string;
-  fullLabel?: string;
-  kind?: "major" | "minor";
-  /** Typed band-domain position; avoids indexing filtered/reordered breaks. */
-  domainIndex?: number;
-  /** false when band-label thinning hides this tick's label. */
-  labeled: boolean;
-  /** Wrapped label lines (band axis, mode "wrapped"). */
-  lines?: string[];
-  /** Rotation in degrees (band axis, mode "rotated"): 0 | -45 | -90. */
-  angle?: number;
-}
-
-export interface AxisResult {
-  ticks: Tick[];
-  guidePlan?: AxisGuidePlan;
-  /** band thinning: only every n-th tick keeps its label. 1 = all. */
-  labelEvery: number;
-  truncated: boolean;
-}
-
-/** Custom formatter: value plus the tick step (NaN for band axes). */
-export type TickFormatter = (value: number | string, step: number) => string;
-
-export interface LayoutTheme {
-  /** Tick label font size in px. */
-  fontSize: number;
-  /** Pass-A margin priors. */
-  marginPriors: Margins;
-  /** Margin floors (axis line + breathing room even with no labels). */
-  minMargins: Margins;
-  tickLength: number;
-  tickLabelGap: number;
-  /** Per-side cap as a fraction of the plot dimension on that side's axis. */
-  maxMarginFraction: number;
-  targetPxPerTickX: number;
-  targetPxPerTickY: number;
-  maxTicks: number;
-  /** Margin quantization grid in px (0 disables). */
-  quantum: number;
-  ellipsis: string;
-}
-
-export const DEFAULT_LAYOUT_THEME: LayoutTheme = {
-  fontSize: 11,
-  marginPriors: { top: 8, right: 12, bottom: 28, left: 44 },
-  minMargins: { top: 4, right: 4, bottom: 16, left: 12 },
-  tickLength: 6,
-  tickLabelGap: 3,
-  maxMarginFraction: 0.35,
-  targetPxPerTickX: 80,
-  targetPxPerTickY: 40,
-  maxTicks: 10,
-  quantum: 4,
-  ellipsis: "…",
-};
-
-export interface LayoutInput {
-  width: number;
-  height: number;
-  x: Domain;
-  y: Domain;
-  formatX?: TickFormatter;
-  formatY?: TickFormatter;
-  measurer: TextMeasurer;
-  theme?: Partial<LayoutTheme>;
-  /**
-   * Extra per-side space consumed by non-tick content (axis titles, legends),
-   * folded INTO the two-pass measurement loop: reserved space shrinks the
-   * panel, which re-derives tick counts — no post-layout margin hacks
-   * (fixes the M0c follow-up in decision 0007).
-   */
-  reserve?: Partial<Margins>;
-  /** @internal Pass-A temporal plans used as monotonic Pass-B hints. */
-  previousGuidePlans?: Readonly<{ x?: AxisGuidePlan; y?: AxisGuidePlan }>;
-}
-
-export interface PassResult {
-  margins: Margins;
-  x: AxisResult;
-  y: AxisResult;
-  degradations: string[];
-}
-
-export interface LayoutResult extends PassResult {
-  innerWidth: number;
-  innerHeight: number;
-  converged: boolean;
-  passes: 2;
-  passAMargins: Margins;
-}
+export {
+  DEFAULT_LAYOUT_THEME,
+  type AxisResult,
+  type BandLayoutDomainContext,
+  type Domain,
+  type LayoutInput,
+  type LayoutResult,
+  type LayoutTheme,
+  type Margins,
+  type PassResult,
+  type TemporalLayoutDomainContext,
+  type Tick,
+  type TickFormatter,
+} from "./layout-types.js";
 
 const clamp = (v: number, lo: number, hi: number) => Math.max(lo, Math.min(hi, v));
 
 function quantizeUp(px: number, quantum: number): number {
   if (quantum <= 0) return px;
   return Math.ceil(px / quantum - 1e-9) * quantum;
-}
-
-interface AxisTicks {
-  ticks: Tick[];
-  step: number;
-  empty: boolean;
-  guidePlan?: AxisGuidePlan;
-  /** Measured band planner: orthogonal (bottom) band height it requires, px. */
-  bandLabelBandHeight?: number;
-  /** Measured band planner: along-axis overhang past the RIGHT end tick, px. */
-  bandAlongOverhang?: number;
-  /** Measured band planner: along-axis overhang past the LEFT end tick, px. */
-  bandLeftOverhang?: number;
-  /** Measured band planner: resolved thinning (1 = every category labeled). */
-  bandLabelEvery?: number;
-}
-
-interface DeriveTicksContext {
-  orient: "horizontal" | "vertical";
-  extentPx: number;
-  measurer: TextMeasurer;
-  fontSize: number;
-  marginCapPx: number;
-  orthogonalMarginCapPx?: number;
-  /** Tick chrome (tickLength + gap) that layoutPass adds AFTER the label band; the
-   *  band planner subtracts it so the band + chrome together honor the margin cap. */
-  orthogonalChromePx?: number;
-  quantum?: number;
-  previousGuidePlan?: AxisGuidePlan;
-}
-
-/** Assemble an AxisGuidePlan from a measured band plan (mirrors planBasicAxis). */
-function bandGuidePlan(
-  plan: BandAxisPlan,
-  context: BandLayoutDomainContext,
-  rawCategories: readonly unknown[],
-): AxisGuidePlan {
-  const reverse = context.config.reverse === true;
-  return Object.freeze({
-    type: "axis" as const,
-    id: `axis:${context.aesthetic}:panel:${String(context.panelIndex)}`,
-    aesthetic: context.aesthetic,
-    panelIndex: context.panelIndex,
-    scaleType: "band" as const,
-    transform: "identity" as const,
-    temporalKind: null,
-    domain: Object.freeze(rawCategories.map((value) => value as CellValue)),
-    direction: reverse ? ("descending" as const) : ("ascending" as const),
-    source: context.config.breaks === undefined ? ("automatic" as const) : ("explicit" as const),
-    interval: null,
-    locale: null,
-    timezone: null,
-    ticks: Object.freeze(
-      plan.ticks.map((tick) =>
-        Object.freeze({
-          value: rawCategories[tick.domainIndex] as CellValue,
-          label: tick.labeled ? tick.label : "",
-          fullLabel: tick.fullLabel,
-          kind: "major" as const,
-        }),
-      ),
-    ),
-    ...(context.config.breaks !== undefined && {
-      sourceBreaks: Object.freeze([...context.config.breaks] as CellValue[]),
-    }),
-    overlap: plan.overlap,
-    marginOverflow: plan.marginOverflow,
-    degraded: Object.freeze([...plan.degraded]),
-    bandLabelMode: plan.mode,
-    bandLabelAngle: plan.angle,
-    bandLabelBandHeight: plan.labelBandHeight,
-  });
-}
-
-function smallestGap(values: readonly number[]): number {
-  let gap = 0;
-  for (let i = 1; i < values.length; i++) {
-    const d = Math.abs(values[i]! - values[i - 1]!);
-    if (d > 0 && (gap === 0 || d < gap)) gap = d;
-  }
-  return gap;
-}
-
-function deriveTicks(
-  domain: Domain,
-  requestedCount: number,
-  format: TickFormatter | undefined,
-  labelEvery: number,
-  context: DeriveTicksContext,
-): AxisTicks {
-  if (domain.type === "band") {
-    const rawCategories = domain.rawCategories ?? domain.categories;
-    // Measured planner: horizontal band axes with a planning context only (G1).
-    // Vertical band (native Y, or categorical-on-Y after coord_flip) falls through
-    // to the legacy thin/truncate path.
-    if (domain.band !== undefined && context.orient === "horizontal") {
-      // Resolve break-filtered entries with the same identity match as the
-      // legacy path (breaks are typed raw values → match rawCategories).
-      const resolved =
-        domain.breaks === undefined
-          ? domain.categories.map((value, domainIndex) => ({ value, domainIndex }))
-          : domain.breaks
-              .map((value) => ({
-                value,
-                domainIndex: rawCategories.findIndex((category) => Object.is(category, value)),
-              }))
-              .filter(({ domainIndex }) => domainIndex >= 0);
-      const plan = planBandAxis({
-        aesthetic: domain.band.aesthetic,
-        panelIndex: domain.band.panelIndex,
-        categoryCount: domain.categories.length,
-        entries: resolved.map(({ value, domainIndex }) => ({
-          value,
-          label: format ? format(value, NaN) : String(value),
-          domainIndex,
-        })),
-        orient: context.orient,
-        extentPx: context.extentPx,
-        reverse: domain.band.config.reverse === true,
-        measurer: context.measurer,
-        fontSize: context.fontSize,
-        marginCapPx: context.marginCapPx,
-        // Reserve tick chrome: layoutPass adds tickLength+tickLabelGap to the band
-        // height and the renderer offsets labels by it, so the label band itself
-        // must fit the cap MINUS that chrome or it clips without truncating.
-        orthogonalMarginCapPx: Math.max(
-          1,
-          (context.orthogonalMarginCapPx ?? context.marginCapPx) -
-            (context.orthogonalChromePx ?? 0),
-        ),
-        ...(context.quantum !== undefined && { quantum: context.quantum }),
-        previousMode: context.previousGuidePlan?.bandLabelMode ?? null,
-      });
-      const ticks: Tick[] = plan.ticks.map((tick) => ({
-        value: tick.value,
-        label: tick.label,
-        fullLabel: tick.fullLabel,
-        domainIndex: tick.domainIndex,
-        labeled: tick.labeled,
-        ...(tick.lines !== undefined && { lines: tick.lines }),
-        ...(tick.angle !== undefined && { angle: tick.angle }),
-      }));
-      return {
-        ticks,
-        step: NaN,
-        empty: plan.ticks.length === 0,
-        guidePlan: bandGuidePlan(plan, domain.band, rawCategories),
-        bandLabelBandHeight: plan.labelBandHeight,
-        bandAlongOverhang: plan.alongOverhang,
-        bandLeftOverhang: plan.leftOverhang,
-        bandLabelEvery: plan.labelEvery,
-      };
-    }
-    const entries =
-      domain.breaks === undefined
-        ? domain.categories.map((value, domainIndex) => ({ value, domainIndex }))
-        : domain.breaks
-            .map((value) => ({
-              value,
-              domainIndex: rawCategories.findIndex((category) => Object.is(category, value)),
-            }))
-            .filter(({ domainIndex }) => domainIndex >= 0);
-    const ticks = entries.map(({ value, domainIndex }, index) => ({
-      value,
-      label: format ? format(value, NaN) : String(value),
-      domainIndex,
-      labeled: index % labelEvery === 0,
-    }));
-    return { ticks, step: NaN, empty: entries.length === 0 };
-  }
-  const { min, max } = domain;
-  if (!Number.isFinite(min) || !Number.isFinite(max)) {
-    return { ticks: [], step: NaN, empty: true };
-  }
-
-  // Explicit breaks override derivation (out-of-domain breaks are dropped).
-  if (domain.breaks !== undefined && !(domain.type === "time" && domain.temporal !== undefined)) {
-    const values = domain.breaks.filter((v) => Number.isFinite(v) && v >= min && v <= max);
-    const step = smallestGap(values);
-    const fmt: (v: number) => string = format
-      ? (v) => format(v, step)
-      : domain.type === "time"
-        ? defaultTimeTickFormat
-        : domain.type === "linear" && domain.transform === "log10"
-          ? defaultLogTickFormat
-          : defaultTickFormat(step);
-    return {
-      ticks: values.map((v) => ({ value: v, label: fmt(v), labeled: true })),
-      step,
-      empty: false,
-    };
-  }
-
-  if (domain.type === "time") {
-    if (domain.temporal !== undefined) {
-      const plan = planTemporalAxis({
-        aesthetic: domain.temporal.aesthetic,
-        panelIndex: domain.temporal.panelIndex,
-        domain: [min, max],
-        kind: domain.temporal.kind,
-        orient: context.orient,
-        extentPx: context.extentPx,
-        reverse: domain.temporal.config.reverse === true,
-        measurer: context.measurer,
-        fontSize: context.fontSize,
-        marginCapPx: context.marginCapPx,
-        ...(context.orthogonalMarginCapPx !== undefined && {
-          orthogonalMarginCapPx: context.orthogonalMarginCapPx,
-        }),
-        config: domain.temporal.config,
-        ...(domain.breaks !== undefined && { breaks: domain.breaks }),
-        ...(domain.temporal.sourceBreaks !== undefined && {
-          sourceBreaks: domain.temporal.sourceBreaks,
-        }),
-        ...(context.previousGuidePlan?.interval !== undefined && {
-          previousInterval: context.previousGuidePlan.interval,
-        }),
-      });
-      return {
-        ticks: plan.ticks.map((tick) => ({
-          value: tick.value as number,
-          label: tick.label,
-          fullLabel: tick.fullLabel,
-          kind: tick.kind,
-          labeled: tick.kind === "major" && tick.label !== "",
-        })),
-        step: NaN,
-        empty: plan.ticks.length === 0,
-        guidePlan: plan,
-      };
-    }
-    const values = timeTicks(min, max, requestedCount).values;
-    const fmt: (v: number) => string = format ? (v) => format(v, NaN) : defaultTimeTickFormat;
-    return {
-      ticks: values.map((v) => ({ value: v, label: fmt(v), labeled: true })),
-      step: NaN,
-      empty: false,
-    };
-  }
-  // log10: decade-aware ticks selected on the semantic (positive) domain.
-  if (domain.transform === "log10") {
-    const values = logTicks(min, max, requestedCount);
-    const fmt: (v: number) => string = format ? (v) => format(v, NaN) : defaultLogTickFormat;
-    return {
-      ticks: values.map((v) => ({ value: v, label: fmt(v), labeled: true })),
-      step: NaN,
-      empty: false,
-    };
-  }
-  // sqrt: select ticks in transformed (sqrt) space, inverse-project to finite
-  // semantic labels so spacing is even in pixels.
-  if (domain.transform === "sqrt") {
-    const lo = min < 0 ? 0 : Math.sqrt(min);
-    const transformedTicks = linearTicks(lo, Math.sqrt(max), requestedCount);
-    const values = transformedTicks.map((t) => t * t);
-    const step = smallestGap(values);
-    const fmt: (v: number) => string = format ? (v) => format(v, step) : defaultTickFormat(step);
-    return {
-      ticks: values.map((v) => ({ value: v, label: fmt(v), labeled: true })),
-      step,
-      empty: values.length === 0,
-    };
-  }
-  const values = linearTicks(min, max, requestedCount);
-  const step = min === max ? 0 : tickStep(min, max, requestedCount);
-  const fmt: (v: number) => string = format ? (v) => format(v, step) : defaultTickFormat(step);
-  return {
-    ticks: values.map((v) => ({ value: v, label: fmt(v), labeled: true })),
-    step,
-    empty: false,
-  };
 }
 
 function maxLabeledWidth(axis: AxisTicks, measurer: TextMeasurer, fontSize: number): number {
